+++ src/main/java/racingcar/model/Cars/Location.java
@@ -1,4 +1,4 @@
-package racingcar.model;
+package racingcar.model.Cars;
 
 public class Location {
