+++ src/main/java/racingcar/model/Race/RaceManager.java
@@ -1,6 +1,8 @@
-package racingcar.model;
+package racingcar.model.Race;
 
 import java.util.List;
+import racingcar.model.Cars.Cars;
+import racingcar.model.Winners.Winners;
 import racingcar.view.InputView;
 import racingcar.view.OutputView;
