+++ src/test/java/racingcar/TryCountTest.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import racingcar.model.TryCount;
+import racingcar.model.Race.TryCount;
 
 public class TryCountTest {
     @DisplayName("입력 받은 시도 횟수가 숫자가 아닌 경우 예외가 발생한다")
